refactor: stop scam request from CollectionIcon
