+++ packages/mask/src/manifest.json
@@ -10,15 +10,6 @@
     "browser_action": { "default_popup": "popups.html" },
     "homepage_url": "https://mask.io",
     "description": "The portal to the new & open Internet. Send encrypted message and decentralized Apps right on top of social networks.",
-    "web_accessible_resources": [
-        "js/*",
-        "*.svg",
-        "*.png",
-        "*.css",
-        "https://gateway.ipfscdn.io/ipfs/*",
-        "https://logo.nftscan.com/*",
-        "https://imagedelivery.net/*",
-        "https://raw.githubusercontent.com/dimensiondev/*"
-    ],
+    "web_accessible_resources": ["js/*", "*.svg", "*.png", "*.css"],
     "content_security_policy": "script-src 'self' 'wasm-eval'; object-src 'self';"
 }
