+++ packages/shared/src/UI/components/Icon/constants.ts
