+++ packages/shared/src/UI/components/TokenIcon/index.tsx
@@ -1,19 +1,12 @@
-import { memo, useState } from 'react'
 import { useAsyncRetry } from 'react-use'
 import { first } from 'lodash-unified'
-import { Avatar, AvatarProps, useTheme } from '@mui/material'
-import { makeStyles, useStylesExtends } from '@masknet/theme'
+import type { AvatarProps } from '@mui/material'
+import { makeStyles } from '@masknet/theme'
 import { useChainId, useWeb3Hub, Web3Helper } from '@masknet/plugin-infra/web3'
 import { NetworkPluginID, TokenType } from '@masknet/web3-shared-base'
 import { EMPTY_LIST } from '@masknet/shared-base'
-import NO_IMAGE_COLOR from './constants'
 import { useImageBase64 } from '../../../hooks/useImageBase64'
-
-const useStyles = makeStyles()((theme) => ({
-    icon: {
-        margin: 0,
-    },
-}))
+import { Icon } from '../Icon'
 
 export interface TokenIconProps extends withClasses<'icon'> {
     chainId?: Web3Helper.ChainIdAll
@@ -50,7 +43,7 @@ export function TokenIcon(props: TokenIconProps) {
     if (!accessibleUrl && originalUrl && disableDefaultIcon) return null
 
     return (
-        <TokenIconUI
+        <Icon
             key={key}
             logoURL={isNFT ? logoURL : accessibleUrl || originalUrl}
             AvatarProps={AvatarProps}
@@ -59,40 +52,3 @@ export function TokenIcon(props: TokenIconProps) {
         />
     )
 }
-
-export interface TokenIconUIProps extends withClasses<'icon'> {
-    logoURL?: string
-    AvatarProps?: Partial<AvatarProps>
-    name?: string
-}
-
-export const TokenIconUI = memo<TokenIconUIProps>((props) => {
-    const { logoURL, AvatarProps, name } = props
-    const [error, setError] = useState(false)
-
-    // add background color to no-img token icon
-    const defaultBackgroundColorNumber = name?.split('')?.reduce((total, cur) => total + Number(cur?.charCodeAt(0)), 0)
-    const defaultBackgroundColor = defaultBackgroundColorNumber
-        ? NO_IMAGE_COLOR[defaultBackgroundColorNumber % 5]
-        : undefined
-    const classes = useStylesExtends(useStyles(), props)
-    const theme = useTheme()
-
-    return (
-        <Avatar
-            className={classes.icon}
-            src={logoURL}
-            {...AvatarProps}
-            imgProps={{
-                onError: () => {
-                    setError(true)
-                },
-            }}
-            sx={{
-                ...AvatarProps?.sx,
-                backgroundColor: logoURL && !error ? theme.palette.common.white : defaultBackgroundColor,
-            }}>
-            {name?.slice(0, 1).toUpperCase()}
-        </Avatar>
-    )
-})
