+++ packages/shared/src/UI/components/index.ts
@@ -28,3 +28,4 @@ export * from './PersonaAction'
 export * from './WalletSettingCard'
 export * from './PluginCardFrame'
 export * from './Markdown'
+export * from './Icon'
